# frontend/contract/swagger.yaml
openapi: 3.1.0
info:
  title: Trailarr API
  summary: Trailarr API available commands.
  description: API for Trailarr application.
  termsOfService: https://github.com/nandyalu/trailarr
  contact:
    name: Trailarr - Github
    url: https://github.com/nandyalu/trailarr
  license:
    name: GNU GPL 3.0
    url: https://github.com/nandyalu/trailarr/blob/main/LICENSE
  version: 0.4.0
servers:
  - url: '{protocol}://{hostpath}'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      hostpath:
        default: 127.0.0.1:7889
paths:
...
  /api/v1/files/files:
    get:
      tags:
        - Files
        - Media
      summary: Get Files
      description: "Get files in a directory.\n\nArgs:\n    path (str): Path to the\
        \ directory.\nReturns:\n    FolderInfo|None: Folder information or None if\
        \ folder doesn't exist. \n\nRaises:\n    HTTPException (404): If the folder\
        \ is not found."
      operationId: get_files_api_v1_files_files_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            title: Path
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/FolderInfo'
                  - type: 'null'
                title: Response Get Files Api V1 Files Files Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/files_simple:
    get:
      tags:
        - Files
        - Media
      summary: Get Files Simple
      description: "Get files in a directory in a simple format.\n\nArgs:\n    path\
        \ (str): Path to the directory. \n\nReturns:\n    list: List of file names\
        \ in the directory. \n\nRaises:\n    HTTPException (404): If the folder is\
        \ not found."
      operationId: get_files_simple_api_v1_files_files_simple_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            title: Path
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FolderInfo'
                title: Response Get Files Simple Api V1 Files Files Simple Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/video:
    get:
      tags:
        - Files
        - Media
      summary: Video Endpoint
      description: "Stream video files.\n\nArgs:\n    file_path (str): Path to the\
        \ video file.\n    range (str, optional=None): Range of bytes to stream. \n\
        \nRaises:\n    HTTPException (400): If the file path is invalid.\n    HTTPException\
        \ (404): If the file is not found.\n    HTTPException (400): If the file is\
        \ not a video file. \n\nReturns:\n    Response: Video stream response."
      operationId: video_endpoint_api_v1_files_video_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: file_path
          in: query
          required: true
          schema:
            type: string
            title: File Path
        - name: range
          in: header
          required: false
          schema:
            type: string
            title: Range
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/read:
    get:
      tags:
        - Files
        - Media
      summary: Read File
      description: Read the contents of a file.
      operationId: read_file_api_v1_files_read_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: file_path
          in: query
          required: true
          schema:
            type: string
            title: File Path
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Read File Api V1 Files Read Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/video_info:
    get:
      tags:
        - Files
        - Media
      summary: Get Video Info
      description: Get information about the video file.
      operationId: get_video_info_api_v1_files_video_info_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: file_path
          in: query
          required: true
          schema:
            type: string
            title: File Path
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/VideoInfo'
                  - type: 'null'
                title: Response Get Video Info Api V1 Files Video Info Get
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/rename:
    post:
      tags:
        - Files
        - Media
      summary: Rename File Fol
      description: Rename a file or folder.
      operationId: rename_file_fol_api_v1_files_rename_post
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: old_path
          in: query
          required: true
          schema:
            type: string
            title: Old Path
        - name: new_path
          in: query
          required: true
          schema:
            type: string
            title: New Path
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: boolean
                title: Response Rename File Fol Api V1 Files Rename Post
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/files/delete:
    delete:
      tags:
        - Files
        - Media
      summary: Delete File Fol
      description: Delete a file/folder from the filesystem.
      operationId: delete_file_fol_api_v1_files_delete_delete
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            title: Path
        - name: media_id
          in: query
          required: false
          schema:
            type: integer
            default: -1
            title: Media Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: boolean
                title: Response Delete File Fol Api V1 Files Delete Delete
        '401':
          description: API key is missing
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/media/{media_id}/files:
    get:
      tags:
        - Media
      summary: Get Media Files
      description: "Get media files by ID. \n\nArgs:\n    media_id (int): ID of the\
        \ media item. \n\nReturns:\n    FolderInfo|str: Folder information or error\
        \ message."
      operationId: get_media_files_api_v1_media__media_id__files_get
      security:
        - APIKeyQuery: []
        - APIKeyHeader: []
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: integer
            title: Media Id
        - name: trailarr_api_key
          in: cookie
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Trailarr Api Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/FolderInfo'
                  - type: string
                title: Response Get Media Files Api V1 Media  Media Id  Files Get
        '401':
          description: API key is missing
        '404':
          description: Media Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
        - message
      title: ErrorResponse
...
    FolderInfo:
      properties:
        type:
          type: string
          title: Type
          description: Type of the entry (file/folder).
          default: folder
        name:
          type: string
          title: Name
          description: Name of the file/folder.
        size:
          type: string
          title: Size
          description: Size of the file/folder in human-readable format (e.g. '10
            KB').
          default: 0 KB
        path:
          type: string
          title: Path
          description: Path of the file/folder.
        files:
          items:
            $ref: '#/components/schemas/FolderInfo'
          type: array
          title: Files
          description: List of FileInfo or FolderInfo objects representing files and
            folders         inside a given folder.
          default: []
        created:
          type: string
          title: Created
          description: Creation time of the folder in 'YYYY-MM-DD HH:MM:SS' format.
      type: object
      required:
        - name
        - path
        - created
      title: FolderInfo
      description: "Contains information about files and (sub)folders in a given folder.\
        \ \n\nAttrs:\n    - type (str): Type of the entry (folder).\n    - name (str):\
        \ Name of the folder.\n    - files (list[FileInfo | FolderInfo]): List of\
        \ FileInfo or FolderInfo\n        objects representing files and folders inside\
        \ a given folder.\n    - created (str): Creation time of the folder in \"\
        YYYY-MM-DD HH:MM:SS\"\n        format."
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StreamInfo:
      properties:
        index:
          type: integer
          title: Index
        codec_type:
          type: string
          title: Codec Type
        codec_name:
          type: string
          title: Codec Name
        coded_height:
          type: integer
          title: Coded Height
          default: 0
        coded_width:
          type: integer
          title: Coded Width
          default: 0
        audio_channels:
          type: integer
          title: Audio Channels
          default: 0
        sample_rate:
          type: integer
          title: Sample Rate
          default: 0
        language:
          type: string
          title: Language
          default: ''
        duration:
          type: string
          title: Duration
          default: ''
      type: object
      required:
        - index
        - codec_type
        - codec_name
      title: StreamInfo
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    VideoInfo:
      properties:
        name:
          type: string
          title: Name
        format_name:
          type: string
          title: Format Name
        duration_seconds:
          type: integer
          title: Duration Seconds
          default: 0
        duration:
          type: string
          title: Duration
          default: 0:00:00
        size:
          type: integer
          title: Size
          default: 0
        bitrate:
          type: string
          title: Bitrate
          default: 0 bps
        streams:
          items:
            $ref: '#/components/schemas/StreamInfo'
          type: array
          title: Streams
      type: object
      required:
        - name
        - format_name
        - streams
      title: VideoInfo
  securitySchemes:
    APIKeyQuery:
      type: apiKey
      in: query
      name: api_key
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-KEY